+++ fortpy/parsers/docstring.py
@@ -1,3 +1,4 @@
+from .. import msg
 import xml.etree.ElementTree as ET
 import re
 from ..docelements import DocElement, DocGroup
@@ -58,7 +59,7 @@ class DocStringParser(object):
                     else:
                         result[key][0].extend(list(docs))
                 except ET.ParseError:
-                    print(doctext)
+                    msg.err(doctext)
         else:
             #This is the text content of a code element that was buried inside of the module.
             #Get all the docblocks and the items they decorate from this parent.
@@ -74,7 +75,7 @@ class DocStringParser(object):
         :arg add: when true, docgroups must be unique in the code element;
           otherwise, existing groups are overwritten."""
         if group.name in code_el.groups and add:
-            print("WARNING: duplicate group names in code element {}".format(code_el.name))
+            msg.warn("duplicate group names in code element {}".format(code_el.name))
         else:
             code_el.groups[group.name] = group
 
@@ -102,9 +103,9 @@ class DocStringParser(object):
                         anexec.parameters[doc.pointsto].overwrite_docs(doc)
                 else: 
                     #the parameter docstring is orphaned, give a warning.
-                    msg = "WARNING: the docstring for parameter '{}' had no corresponding " + \
+                    wmsg = "the docstring for parameter '{}' had no corresponding " + \
                           "parameter in the executable definition for '{}'."
-                    print(msg.format(doc.pointsto, anexec))
+                    msg.warn(wmsg.format(doc.pointsto, anexec))
             elif doc.doctype == "group":
                 kids = self._process_docgroup(doc, anexec)
                 if add:
@@ -260,7 +261,7 @@ class DocStringParser(object):
                         else:
                             docblocks[key][0].extend(list(docs))
                     except ET.ParseError:
-                        print(doctext)
+                        msg.err(doctext)
 
                     #Reset the list of current docstrings
                     current = []
@@ -328,10 +329,10 @@ class DocStringParser(object):
                     docs = self.to_doc(xmldict[kdecor], memname)
                     self.process_execdocs(docs, member, kdecor)
                 else:
-                    print("WARNING: orphaned docstring. No member {} in module {}.".format(
+                    msg.warn("orphaned docstring. No member {} in module {}.".format(
                         member, modname))
             else:
-                print("WARNING: orphaned docstring from XML docfile for {}".format(kdecor))
+                msg.warn("orphaned docstring from XML docfile for {}".format(kdecor))
 
     def rt_update_module(self, xmldict, module):
         """Updates the members, executables and types in the specified module