+++ fortpy/testing/comparer.py
@@ -1,3 +1,4 @@
+from .. import msg
 import xml.etree.ElementTree as ET
 import re
 import os
@@ -171,8 +172,8 @@ class FileRepresentation(object):
             if line.count in store:
                 count = store[line.count]
             else:
-                print("\nFATAL: reference to stored value that does not exist '{}'".format(line.count))
-                print("STORE: {}".format(list(store.keys())))
+                msg.err("reference to stored value that does not exist '{}'".format(line.count))
+                msg.gen("STORE: {}".format(list(store.keys())))
                 exit(1)
 
         return count
@@ -230,15 +231,15 @@ def compare_representations(rep1, rep2, mode = "default"):
 
         return result
     else:
-        print("FATAL: a templated representation cannot be compared to a simple one.")
+        msg.err("a templated representation cannot be compared to a simple one.")
         
 def _compare_validate(rep, mode):
     """Validates the contents of the representation."""
     if not mode in rep.template.comparisons:
-        print("FATAL: the mode '{}' specified for comparisons does not exist.".format(mode))
+        msg.err("the mode '{}' specified for comparisons does not exist.".format(mode))
         exit(1)
     if not mode in rep.template.outcomes:
-        print("FATAL: the mode '{}' specified for outcomes does not exist.".format(mode))
+        msg.err("the mode '{}' specified for outcomes does not exist.".format(mode))
         exit(1)
 
 def _compare_preamble(rep1, rep2, mode):
@@ -323,8 +324,9 @@ def _compare_block(rep1, rep2, thisb, thatb, mode, bkey = None, index = None, is
         if thatv is not None:
             if len(thisv) != len(thatv):
                 loop = min([len(thisv), len(thatv)])
-                print("""WARNING: line values extracted for key '{}' in the two files
-have different numbers of elements: {} vs. {}.""".format(key, len(thisv), len(thatv)))
+                msg.warn("line values extracted for key" + 
+                         " '{}' in the two files have different".format(key) + 
+                         " numbers of elements: {} vs. {}.".format(len(thisv), len(thatv)))
             else:
                 loop = len(thisv)
 
@@ -450,7 +452,7 @@ def _get_key_value_single(key, bodyblock):
             result = bodyblock[key].values[0]
 
     if result is None:
-        print("FATAL: block {} did not return a valid key value for '{}'.".format(bodyblock, template.key))
+        msg.err("block {} did not return a valid key value for '{}'.".format(bodyblock, template.key))
         exit(1)
 
     return result    
@@ -497,8 +499,8 @@ class FileComparer(object):
         """
         source = os.path.expanduser(path)
         if not os.path.exists(source):
-            print("ERROR: can't create representation for {}.".format(source) + 
-                  " File does not exist.\n")            
+            msg.err("can't create representation for {}.".format(source) + 
+                    " File does not exist.\n")            
             return None
 
         with open(source) as f:
@@ -527,7 +529,7 @@ class FileComparer(object):
             svalues = FileRepresentation(slines, stemplate, sv, source)
 
         if not svalues.extracted:
-            print("\nERROR: ouput file does not have the same format as template.\n{}".format(source))
+            msg.err("ouput file does not have the same format as template.\n{}".format(source))
             return None
         else:
             return svalues
@@ -549,10 +551,10 @@ class FileComparer(object):
                 else:
                     return None
             except ET.ParseError:
-                print("WARNING: no version information found in the file. Assuming version 1.")
+                msg.warn("no version information found in the file. Assuming version 1.")
                 return None
         else:
-            print("WARNING: no version information found in the file. Assuming version 1.")
+            msg.warn("no version information found in the file. Assuming version 1.")
             return None
 
     def _get_file_version(self, fortpyxml):