+++ fortpy/testing/executable.py
@@ -1,3 +1,4 @@
+from .. import msg
 from fortpy.testing.method import MethodWriter
 from os import path, mkdir, remove
 from datetime import datetime
@@ -66,7 +67,7 @@ class ExecutableGenerator(object):
                 newi = result.index(modn)
                 self._process_module_needs(modn, newi, result)
         else:
-            print("FATAL: unable to find module {}.".format(module))
+            msg.err("unable to find module {}.".format(module))
             exit(1)
 
     def reset(self, identifier, libraryroot, rerun = False):
@@ -80,7 +81,7 @@ class ExecutableGenerator(object):
 
         #Create the directory for the executable files to be copied and written to.
         if not path.exists(self.folder):
-            print("EXEC DIR: create {}".format(self.folder))
+            msg.okay("EXEC DIR: create {}".format(self.folder))
             mkdir(self.folder)
 
         #If re-run is specified, delete the fortpy.f90 file to force